docker/requirements.txt
streamlit==1.34.0
farm-haystack==1.26.2
elasticsearch==8.14.0
nltk==3.8.1
langdetect==1.0.9